renames remote module to decorators
