+++ tests/test_remote.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 import pytest
 
 from block_cascade import GcpEnvironmentConfig, GcpMachineConfig, GcpResource
-from block_cascade.remote import remote
+from block_cascade import remote
 from block_cascade.utils import PREFECT_VERSION
 
 if PREFECT_VERSION == 2:
